02: solve part2
This was a pretty simple adjustment of the code for the first bit.
Now simply the returned value pair no longer indicates lower and
upper bound but instead indicates the two indices at which we expect
to find letter, but only find it once between the two of them!

We no longer have to std::count the letter occurrences and instead
now simply XOR whether at either place there is the letter.