11: solve part 1
Fairly easy, but I had a lingering bug in my iteration inner loop,
where I checked for _both_ rows and columns that the index was
within the number of _rows_.
This didn't make a difference for the 10x10 sample input, but did
for the actual one.
I'm embarassed that even though I wrote some tests to debug this, I
didn't think of testing this fairly obvious case for a long time.