14: solve part 1 in C++
Learned something new while tackling an overflow issue on this one.
The 64 bit unsigned was definitely big enough here, but stuff kept
getting truncated to 32 bits somewhere.
Turns out, the type inference (or whatever C++ heads call it)
doesn't use lambda parameter to determine the input type for a
`std::accumulate` parameter, but instead you're required to cast it
to the correct type.
So it needs to be `std::accumulate(..., (uint64_t)0, ...)` for it
to actually work.
Very odd that this didn't even warn about a type mismatch or
implicit narrowing cast or something, because I did explicitly
return a 64 bit value but then it would apparently get used as a 32
bit input value right after.