15: solve part 1 naively in C++
Pretty simple, this is a nice task for Haskell for example, so I
tried doing it in the same simple way in C++, using std::find.
This is the naive approach, I think the smart approach (which seems
will be needed for part 2) would be to keep a map associating each
number with the iterator of its most recent occurrence.