20: Do part 1
Very enjoyable but I don't like that the solution is a bit of a
cop-out, because I only check how many matches exist for each piece
and then grab four pieces that only have 2 matches and call them the
corner pieces. It works for the inputs we're given, but it's not a
great general solution and I'd like to at least improve it.