18: Solve in Go
The implementation itself is pretty boring and samey, but I was
curious about how to do the
  Expression      := BinaryOperation | Value
  BinaryOperation := Expression Operation Expression
  Operation       := Add | Multiply
thing in Go.

(I definitely didn't know what to google to figure out how to model
 this at first.)

The missing link for me was how interfaces worked for this; it's
pretty neat:
 - The `Expression` interface defines `eval()`
 - The `BinaryOperation` struct has two `*Expression`s (and an Op)
 - The `ValueExpr` struct just has a value

When I let the parser parse, it returns an `Expression`. When I call
`eval()` on that, it calls the right functions all the way down.

Lots left to learn about this.